Se corrigio el calculo incorrecto producido por operaciones con un tipo de dato int y otro conun tipo de dato double.

+++ AlcoholimetroSSP/Domain/IopeSegPub.cs
@@ -5,7 +5,7 @@ namespace AlcoholimetroSSP.Domain
 {
     public interface IopeSegPub
     {
-        double calcular (string bebida, int cantidad, double peso);
+        double calcular (string bebida, double cantidad, double peso);
 
     }
 }
